requirements.txt
allure-pytest==2.9.44
allure-python-commons==2.9.44
decorator==4.4.1
pytest~=6.2.4
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-xdist==2.3.0
selenium==3.141.0
webdriver-manager==3.4.2
python-dotenv~=0.19.0
pylint==2.12.2
